chore(terraform-provider): refactor terraform manager
