+++ lib/manager/terraform/extract.spec.ts
@@ -1,9 +1,5 @@
 import { readFileSync } from 'fs';
-import {
-  TerraformDependencyTypes,
-  extractPackageFile,
-  getTerraformDependencyType,
-} from './extract';
+import { extractPackageFile } from './extract';
 
 const tf1 = readFileSync('lib/manager/terraform/__fixtures__/1.tf', 'utf8');
 const tf2 = `module "relative" {
@@ -24,38 +20,6 @@ describe('lib/manager/terraform/extract', () => {
     });
     it('returns null if only local deps', () => {
       expect(extractPackageFile(tf2)).toBeNull();
-    });
-  });
-  describe('getTerraformDependencyType()', () => {
-    it('returns TerraformDependencyTypes.module', () => {
-      expect(getTerraformDependencyType('module')).toBe(
-        TerraformDependencyTypes.module
-      );
-    });
-    it('returns TerraformDependencyTypes.provider', () => {
-      expect(getTerraformDependencyType('provider')).toBe(
-        TerraformDependencyTypes.provider
-      );
-    });
-    it('returns TerraformDependencyTypes.unknown', () => {
-      expect(getTerraformDependencyType('unknown')).toBe(
-        TerraformDependencyTypes.unknown
-      );
-    });
-    it('returns TerraformDependencyTypes.required_providers', () => {
-      expect(getTerraformDependencyType('required_providers')).toBe(
-        TerraformDependencyTypes.required_providers
-      );
-    });
-    it('returns TerraformDependencyTypes.unknown on empty string', () => {
-      expect(getTerraformDependencyType('')).toBe(
-        TerraformDependencyTypes.unknown
-      );
-    });
-    it('returns TerraformDependencyTypes.unknown on string with random chars', () => {
-      expect(getTerraformDependencyType('sdfsgdsfadfhfghfhgdfsdf')).toBe(
-        TerraformDependencyTypes.unknown
-      );
     });
   });
 });
